refactor(nvim): miniディレクトリを表示・操作・探索の3ファイルに再編
- ui.lua: 表示系 (statusline, notify, hipatterns, icons, indentscope, diff)
- editor.lua: 操作系 (completion, snippets, clue, pairs, surround, move, bracketed, jump2d)
- nav.lua: 探索系 (files, pick, git, misc, trailspace, basics, starter)
- init.lua: 3ファイルをrequireするだけのエントリポイントに簡素化
