--- config/nvim/lua/my/mini/files.lua
@@ -1,13 +1,0 @@
-MiniDeps.now(function()
-  require('mini.files').setup()
-
-  vim.api.nvim_create_user_command(
-    'Files',
-    function()
-      MiniFiles.open()
-    end,
-    { desc = 'Open file exproler' }
-  )
-  -- command+bでFilesを実行する
-  vim.keymap.set('n', '<D-b>', ':Files<CR>', { silent = true, desc = 'Toggle files' })
-end)
