--- config/nvim/lua/my/mini/git.lua
@@ -1,5 +1,0 @@
-MiniDeps.later(function()
-  require('mini.git').setup()
-
-  vim.keymap.set({ 'n', 'x' }, '<space>gs', MiniGit.show_at_cursor, { desc = 'Show at cursor' })
-end)
