--- config/nvim/lua/my/mini/hipatterns.lua
@@ -1,13 +1,0 @@
--- specific highlight pattern
-MiniDeps.later(function()
-  local hipatterns = require('mini.hipatterns')
-  local hi_words = require('mini.extra').gen_highlighter.words
-  hipatterns.setup({
-    highlighters = {
-      -- Highlight standalone 'TODO'
-      todo = hi_words({ 'TODO' }, 'MiniHipatternsTodo'),
-      -- Highlight hex color strings (`#rrggbb`) using that color
-      hex_color = hipatterns.gen_highlighter.hex_color(),
-    },
-  })
-end)
