+++ config/nvim/lua/my/mini/init.lua
@@ -1,25 +1,5 @@
--- Simple plugins (now)
-for _, name in ipairs({ 'icons', 'basics', 'starter' }) do
-  MiniDeps.now(require('mini.' .. name).setup)
-end
-
--- Simple plugins (later)
-for _, name in ipairs({ 'indentscope', 'pairs', 'diff', 'jump2d', 'bracketed', 'surround', 'move' }) do
-  MiniDeps.later(require('mini.' .. name).setup)
-end
-
 vim.o.winbar = " %m %f "
 
--- Plugins with custom settings
-require('my.mini.statusline')
-require('my.mini.notify')
-require('my.mini.misc')
-require('my.mini.hipatterns')
-require('my.mini.files')
-require('my.mini.pick')
-require('my.mini.completion')
-require('my.mini.snippets')
-require('my.mini.trailspace')
-require('my.mini.git')
-require('my.mini.clue')
-
+require('my.mini.ui')
+require('my.mini.editor')
+require('my.mini.nav')
