--- config/nvim/lua/my/mini/notify.lua
@@ -1,5 +1,0 @@
-MiniDeps.now(function()
-  require('mini.notify').setup()
-
-  vim.notify = require('mini.notify').make_notify({})
-end)
