--- config/nvim/lua/my/mini/pick.lua
@@ -1,9 +1,0 @@
-MiniDeps.later(function()
-  require('mini.pick').setup()
-
-  vim.ui.select = MiniPick.ui_select
-
-  vim.keymap.set('n', '<space>f', function()
-    MiniPick.builtin.files({ tool = 'git' })
-  end, { desc = 'mini.pick.files' })
-end)
