--- config/nvim/lua/my/mini/statusline.lua
@@ -1,5 +1,0 @@
-MiniDeps.now(function()
-  require('mini.statusline').setup()
-  vim.opt.laststatus = 3
-  vim.opt.cmdheight = 0
-end)
