--- config/nvim/lua/my/mini/trailspace.lua
@@ -1,11 +1,0 @@
-MiniDeps.later(function()
-  require('mini.trailspace').setup()
-  vim.api.nvim_create_user_command(
-    'Trim',
-    function()
-      MiniTrailspace.trim()
-      MiniTrailspace.trim_last_lines()
-    end,
-    { desc = 'Trim trailing space and last blank lines' }
-  )
-end)
